requirements.txt
streamlit==1.45.0
groq==0.13.0
openai-whisper==20240930
streamlit-mic-recorder==0.0.8
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.2.5
google-generativeai==0.8.3
ffmpeg-python==0.2.0
